refactor(rename_models): changed file typo
