+++ app/content/models/__init__.py
@@ -7,7 +7,7 @@ from app.content.models.user import User, UserManager
 from app.content.models.user_badge import UserBadge
 from app.content.models.registration import Registration
 from app.content.models.page import Page
-from app.content.models.prioritiy import Priority
+from app.content.models.priority import Priority
 from app.content.models.short_link import ShortLink
 from app.content.models.strike import (
     Strike,
